Module mc_eth_client::config

Base Ethereum client configuration.

Use it as is or reuse top-level fields in your config (“inherit”) in order to share the same configuration file.

struct MyServiceConfig { pub provider: EthereumProviderConfig, pub wallet: Option, pub contracts: Option, }

Default provider and wallet configurations are set for use with Anvil

local Ethereum environment.

Structs

EthereumClientConfig

HttpProviderConfig

LocalWalletConfig

StarknetContracts

Enums

EthereumProviderConfig

EthereumWalletConfig

Constants

DEFAULT_CHAIN_ID

Default Anvil chain ID

DEFAULT_PRIVATE_KEY

Default private key derived from starting Anvil as follows: anvil -b 5 –config-out $BUILD_DIR/anvil.json PRE_PRIVATE=$(jq -r ‘.private_keys[0]’ $BUILD_DIR/anvil.json)

DEFAULT_RPC_ENDPOINT

Default Anvil local endpoint